:wrench: Disable additional linter rule
The Morfologik rule was disabled in VS Code settings to avoid unnecessary warnings.

:art: Add base styles CSS

The styles.css file was given some base universal and functional CSS classes/selectors.

:card_index: Initialize home and contact pages

The index.html and contact.html pages were initialized with boilerplate HTML and linked to the styles stylesheet. This lays the foundation for the lab pages.

Signed-off-by: Zliced13 <07e5297d5b@c0x0.com>