:art: Add logo to horizontal oval
The site logo text was added inside the horizontal oval div for visibility.

:pencil: Refactor CSS comments

The CSS comments were refactored to be more descriptive of the intended styles and group related rules.

:lipstick: Adjust logo styling

Further tweaks were made to styling of the horizontal oval containing the logo, like sizing, positioning and flex formatting.

:policeman: Enforce styles

Some styles were adjusted and the !important flag added in a few places to properly enforce the intended styling overrides.

Signed-off-by: Zliced13 <07e5297d5b@c0x0.com>