:wrench: Remove button left offset
The positioning of the first button was removed since it was not needed.

:pencil: Refine line under header

Further tweaks were made to the styling of the underline element, removing unnecessary margins/widths.

:art: Style buttons

Basic styles were added to format the buttons as block elements and remove border radius.

:lipstick: Group related selectors

The CSS was reorganized to group related selectors together like button styles and ID selectors for better readability.

Signed-off-by: Zliced13 <07e5297d5b@c0x0.com>